==> scripts/1_clean_data.qmd <==
---
title: "Clean Data"
author: "Andreas Eich"
format: html
editor: visual
---

### Libraries

```{r}
library(here) # easier paths, here() function starts in the folder of the RStudio project
library(tidyverse)
library(readxl) # read Excel files
library(janitor) # clean up column names
```

### Read data

```{r}
dat <- read_excel(here("data/raw/2024_09_17_Ashmorebehaviour_new.xlsx")) %>% 
  clean_names()
```

### Clean data

-   Filter by families

-   combine `habitat` levels

-   reorder factor levels for `size_class`

-   filter columns

```{r}
dat <- dat %>% 
  filter(family %in% c("Lutjanidae", "Carangidae", "Serranidae", "Lethrinidae")) %>% 
  mutate(year = factor(year),
         size_class = fct_relevel(size_class, c("<50", "50-100", ">100")),
         habitat = recode(habitat, 
                          "Sand" = "near reef",
                          "Other" = "near reef",
                          "Reef" = "reef",
                          )) %>% 
  filter(!is.na(size_class)) %>% 
  select(video,
         year,
         depth,
         habitat,
         treatment,
         genus_species,
         size_class,
         passing_time_sec,
         feeding_time_sec,
         investigating_time_sec, 
         circling_time_sec,
         bout,
         bite,
         total_time_shark_present_frame,
         n_frame_total)
```

Divide in data for absent/after analysis

-   change order of levels for `treatment`

Select only `absent` and `after`

```{r}
dat <- dat %>% 
  filter(treatment != "before") %>% 
  droplevels()
```

Recode `size_class`

```{r}
dat <- dat %>% 
  mutate(size_class = recode(
    size_class,
    "<50" = "Small",
    "50-100" = "Medium",
    ">100" = "Large"
  )) %>% 
  mutate(size_class = fct_relevel(size_class, c("Small", "Medium", "Large")))
```

Relabel treatments to `absent` and `present` , rename to `shark_in_video`

```{r}
dat <- dat %>% 
  mutate(treatment = recode(treatment, "after" = "present")) %>% 
  rename("shark_in_video" = "treatment")
```

Combine investigating and circling to `investigating_circling`

```{r}
dat <- dat %>% 
  mutate(sec_investigating_circling = investigating_time_sec + circling_time_sec) %>% 
  select(-investigating_time_sec, -circling_time_sec)

```

Rename columns

```{r}
dat <- dat %>% 
  rename("sec_feeding" = "feeding_time_sec",
         "sec_passing" = "passing_time_sec")
```

Calculate total observation times and fractions of behavior

```{r}
dat <- dat %>% 
  mutate(sec_total = sec_feeding + sec_passing + sec_investigating_circling,
         frac_feeding = sec_feeding / sec_total,
         frac_passing = sec_passing / sec_total,
         frac_investigating_circling = sec_investigating_circling / sec_total)
```

Year to factor

```{r}
dat <- dat %>% 
  mutate(year = factor(year))
```

Calculate `log_depth`, `log_n_frame_total`, and `log_total_time_shark_present_frame`

```{r}
dat <- dat %>% 
  mutate(log_depth = log(depth),
         log_n_frame_total = log(n_frame_total),
         log_total_time_shark_present_frame = log(total_time_shark_present_frame))
```

Reorder

```{r}
dat <- dat %>% 
  select(genus_species,
         size_class,
         year,
         shark_in_video,
         sec_investigating_circling,
         sec_passing,
         sec_feeding,
         sec_total,
         frac_investigating_circling,
         frac_passing,
         frac_feeding,
         bout,
         bite,
         habitat,
         depth,
         log_depth,
         total_time_shark_present_frame,
         log_total_time_shark_present_frame,
         n_frame_total,
         log_n_frame_total,
         video)
```

Write data as `.csv` to share and `.rds` to keep formatting

```{r}
write.csv(dat, file = here("data/clean/clean_data.csv"), row.names = F)
write_rds(dat, file = here("data/clean/clean_data.rds"))
```
