==> scripts/2_mod_perc_before_after.qmd <==
---
title: "Model behaviour Percent Before <> After"
author: "Andreas Eich"
format: html
editor: visual
bibliography: references.bib
---

### Libraries

```{r}
library(here) # easier paths, here() function starts in the folder of the RStudio project
library(tidyverse)
library(brms)
library(tidybayes)
library(ggtern)
library(kableExtra)

# get custom functions
source(here("scripts/functions.R"))
```

### Get data

```{r}
dat <- readRDS(here("data/clean/dat_before_after.rds"))
```

-   Ccombine circling and investigating, calculate relative time of behavior types,

-   Make long and wide data frame

```{r}
dat_percW <- dat %>%
  mutate(investigating_circling_time_sec = circling_time_sec + investigating_time_sec,
         total_obs_sec = investigating_circling_time_sec + passing_time_sec + feeding_time_sec,
         investigating_circling_frac = investigating_circling_time_sec/total_obs_sec,
         feeding_frac = feeding_time_sec/total_obs_sec,
         passing_frac = passing_time_sec/total_obs_sec)
         

dat_percL <- dat_percW %>%
  pivot_longer(c(investigating_circling_frac, 
                 feeding_frac, 
                 passing_frac), 
               names_to = "type", 
               values_to = "perc")
```

### Overviews

#### Feeding

Table with number of observations, time spent feeding, per size class, year, and treatment

```{r}
left_join(
  # combine 2 data frames:
  # Part 1: All data for total number of observations,, mean and cumulative feeding time, and mean percent tiem spent feeding
  dat_percW %>%
    group_by(size_class, year, treatment) %>%
    summarise(
      n_total = n(),
      sec_feeding = sum(feeding_time_sec, na.rm = T),
      mean_feeding_sec = mean(feeding_time_sec, na.rm = T) %>%
        round(2),
      mean_feeding_perc = mean(feeding_frac * 100, na.rm = T) %>%
        round(2)
    ),
  
  # Part 2: Filtered for feeding observations to get number of feeding observations
  dat_percW %>%
    filter(feeding_time_sec > 0) %>% #filter
    group_by(size_class, year, treatment) %>%
    summarise(n_feeding = n()),
  by = c("size_class", "year", "treatment")
  
) %>%
  #select relevant columns
  select(treatment,
         n_total,
         n_feeding,
         sec_feeding,
         mean_feeding_sec,
         mean_feeding_perc) %>%
  #replace NAs with 0s
  replace(is.na(.), 0) %>%
  ungroup() %>%
  # group HTML table by size classes
  select(-size_class) %>%
  nice_table() %>%
  pack_rows("<50", 1, 4) %>%
  pack_rows("50-100", 5, 8) %>%
  pack_rows(">100", 9, 12)
```

-   very low number ob observations for large size class \> separate model

#### Videos

Table with number of videos per year. Should be 12

```{r}
dat_percW %>% 
  select(year, video) %>% 
  unique() %>% 
  group_by(year) %>% 
  summarise(n_vid = n()) %>% 
  nice_table()
```

Only 11 videos. Verified with original data:

-   2004: one video without fish observations (`ASH1_1`)

-   2016: one video only treatment = absent, i.e. not in this df

    **CHECK!!! More video missing??**

Table videos per size class

```{r}
dat_percW %>% 
  select(size_class, year, video) %>% 
  unique() %>% 
  group_by(size_class, year) %>% 
  summarise(n_vid = n()) %>% 
  nice_table()
```

### Plot raw data

2004

```{r}
dat_percL %>% 
  filter(year == 2004) %>% 
  group_by(type, treatment, size_class, year) %>% 
  summarise(perc_mean = mean(perc, na.rm = T),
            sd = sd(perc, na.rm = T),
            n = n(),
            se = ifelse(n >=3, sd / sqrt(n), NA)) %>% 
  ggplot()+
  geom_errorbar(aes(x = treatment, 
                    ymin = perc_mean - se, 
                    ymax = perc_mean + se, 
                    col = factor(treatment),
                    width = .4))+
  geom_point(aes(x = treatment, 
                    y = perc_mean, 
                    col = factor(treatment)))+
  facet_grid(type ~size_class , scales = "free")+
  ggtitle("2004")+
  theme_clean()
```

2016

```{r}
dat_percL %>% 
  filter(year == 2016) %>% 
  group_by(type, treatment, size_class, year) %>% 
  summarise(perc_mean = mean(perc, na.rm = T),
            sd = sd(perc, na.rm = T),
            n = n(),
            se = ifelse(n >=3, sd / sqrt(n), NA)) %>% 
  ggplot()+
  geom_errorbar(aes(x = treatment, 
                    ymin = perc_mean - se, 
                    ymax = perc_mean + se, 
                    col = factor(treatment),
                    width = .4))+
  geom_point(aes(x = treatment, 
                    y = perc_mean, 
                    col = factor(treatment)))+
  facet_grid(type ~size_class , scales = "free")+
  ggtitle("2016")+
  theme_clean()
```

As Tenary plot

Small

```{r}
dat_percW %>%
  filter(size_class ==  "<50") %>%
  ggtern(aes(x = investigating_circling_frac, y = passing_frac, z = feeding_frac)) +
  geom_point(aes(col = year))+
  facet_grid(~treatment)+
  theme_tern_clean()
```

Medium

```{r}
dat_percW %>%
  filter(size_class ==  "50-100") %>%
  ggtern(aes(x = investigating_circling_frac, y = passing_frac, z = feeding_frac)) +
  geom_point(aes(col = year))+
  facet_grid(~treatment)+
  theme_tern_clean()
```

Large

```{r}
dat_percW %>%
  filter(size_class ==  ">100") %>%
  ggtern(aes(x = investigating_circling_frac, y = passing_frac, z = feeding_frac)) +
  geom_point(aes(col = year))+
  facet_grid(~treatment)+
  theme_tern_clean()
```

### Model

#### Prepare data

-   Transform according to @douma2019

```{r}
dat_model <- dat_percW %>% 
mutate(frac_circ_invT = transformD(investigating_circling_frac),
         frac_passingT = transformD(passing_frac),
         frac_feedingT = transformD(feeding_frac),
         
         frac_tot = frac_circ_invT + frac_passingT + frac_feedingT,
         
         frac_circ_invT = frac_circ_invT/frac_tot,
         frac_passingT = frac_passingT/frac_tot,
         frac_feedingT = frac_feedingT/frac_tot
         )
```

Calculate logged values for continuous explanatory variables

```{r}
dat_model <- dat_model %>% 
  mutate(log_n_frame_total = log(n_frame_total),
         log_depth = log(depth))
```

#### Small & Medium Fish

Fit model Small an Medium size class

-   use default smooth for now

-   smooth only for response, not phi

-   random intercept only for response, not phi

```{r}
fit_perc_SM <- brm(bf(cbind(frac_circ_invT, frac_passingT, frac_feedingT) ~ 
                factor(year)*treatment*size_class + 
                  s(log_n_frame_total) + 
                  habitat + 
                  s(log_depth) +
                  (1|video),
                phi ~ factor(year)*treatment*size_class),
           data = dat_model %>% 
             filter(size_class != ">100"),
           family =  dirichlet(),
           control = list(max_treedepth = 12, 
                          adapt_delta = 0.98),
           iter = 4000,
           warmup = 2000,
           chains = 4,
           cores = 4,
           seed = 123,
           file_refit = "on_change",
           threads = threading(2, static = T),
           file = here("models/fit_perc_SM"),
           backend = "cmdstanr")
```

```{r}
summary(fit_perc_SM)#1 divergent transition bit Rhat OK
#plot(fit_perc_SM) # ok
```

Get expected values

```{r}
epred_perc_SM <-  epred_draws(
  fit_perc_SM,
  newdata = expand_grid(year = dat_model$year %>% unique(),
                        treatment = dat_model$treatment %>% unique(),
                        size_class = c("<50", "50-100"),
                        habitat = dat_model$habitat %>% unique(),
                        log_depth = dat_model$log_depth %>% mean(),
                        log_n_frame_total = mean(dat_model$n_frame_total) %>% 
                          log()),
  re_formula = NA)
```

Model Plots

```{r}
cols_treat <- c("#FFDB6D", "#4E84C4")

epred_perc_SM %>%
  ggplot(aes(x = .epred, y = year)) +
  stat_slab(size = .3,
            aes(fill = treatment),
            slab_size = 0.5, 
            alpha = 0.3) +
  stat_slab(show.legend = F, 
            size = .3,
            fill = NA, 
            aes(colour = treatment), 
            slab_size = 0.5) +
  scale_colour_manual(values = cols_treat)+
  scale_fill_manual(values = cols_treat)+
  guides(color = guide_legend(override.aes = list(alpha = 1)),
         fill = guide_legend(override.aes = list(alpha = 1)))+
  facet_grid(size_class ~ .category)+
  theme_clean()
```

As tenary plot. white lines are 89% CIs

**Compare before after in one plot**

```{r}
epred_perc_SM %>%
  pivot_wider(names_from = .category, values_from = .epred) %>%
  ggtern(aes(x = frac_circ_invT, y = frac_passingT, z = frac_feedingT)) +
  geom_point(aes(col = treatment), alpha = .1, shape = 1) +
  stat_confidence_tern(
    aes(group = treatment),
    col = "white",
    breaks = c(0.89),
    show.legend = F
  ) +
  labs(
    x = "",
    xarrow  = "Circling/Investigating",
    y = "",
    yarrow  = "Passing",
    z = "",
    zarrow  = "Feeding"
  ) +
  scale_colour_manual(values = cols_treat)+
  guides(color = guide_legend(override.aes = list(alpha = 1)))+
  facet_grid(size_class ~ year) +
  theme_tern_clean()
```

**Compare years in one plot**

```{r}

cols_year <- c("#52854C", "#D16103")

epred_perc_SM %>%
  pivot_wider(names_from = .category, values_from = .epred) %>%
  ggtern(aes(x = frac_circ_invT, y = frac_passingT, z = frac_feedingT)) +
  geom_point(aes(col = year), alpha = .1, shape = 1) +
  stat_confidence_tern(
    aes(group = year),
    col = "white",
    breaks = c(0.89),
    show.legend = F
  ) +
  scale_colour_manual(values = cols_year) +
  guides(color = guide_legend(override.aes = list(alpha = 1))) +
  facet_grid(size_class ~ treatment) +
  theme_tern_clean()
```

Estimated value for difference between treatments. If positive, more time spent with respective behavior after shark appeared.

```{r}
epred_perc_effect_SM <- epred_perc_SM %>%
  ungroup() %>% 
  dplyr::select(year, treatment, .draw, .epred,.category, size_class, habitat, log_n_frame_total, log_depth) %>%
  pivot_wider(names_from = treatment, values_from = .epred) %>%
  mutate(diff = after - before)
 


epred_perc_effect_SM %>% 
  ggplot(aes(x = diff, y = year)) +
  geom_vline(xintercept = 0, linetype = "11", col = "grey")+
  stat_slab(alpha = 0.3, aes(fill = factor(year))) +
  stat_slab(show.legend = F,size = .5,  fill = NA ,  aes(col = factor(year))) +
  scale_colour_manual(values = cols_year) +
  scale_fill_manual(values = cols_year) +
  guides(color = guide_legend(override.aes = list(alpha = 1)),
         fill = guide_legend(override.aes = list(alpha = 1))) +
  facet_grid(size_class ~ .category, scales = "free")+
  theme_clean()
```

Compare distribution of expected values of effect to 0

```{r}
epred_perc_effect_SM %>% 
  compare_0(variable = diff,
          size_class, year, .category)
```

Impact of depth

```{r}
epred_perc_SM_depth <-  epred_draws(
  fit_perc_SM,
  newdata = expand_grid(year = dat_model$year %>% unique(),
                        treatment = dat_model$treatment %>% unique(),
                        size_class = c("<50", "50-100"),
                        habitat = dat_model$habitat %>% unique(),
                        log_depth = seq(min(dat_model$depth),
                                        max(dat_model$depth),
                                        length = 100) %>% 
                          log(),
                        log_n_frame_total = mean(dat_model$n_frame_total) %>% 
                          log()),
  re_formula = NA)


epred_perc_SM_depthS <- epred_perc_SM_depth %>% 
  group_by(.category, log_depth) %>% 
  summarise(median = median(.epred),
            l_66 = quantile(.epred, probs = (1-0.66)/2),
            u_66 = quantile(.epred, probs = 1 - (1-0.66)/2),
            l_89 = quantile(.epred, probs = (1-0.89)/2),
            u_89 = quantile(.epred, probs = 1-(1-0.89)/2),
            l95 = quantile(.epred, 0.025, na.rm = T),
            u95 = quantile(.epred, 0.975, na.rm = T),
      )

epred_perc_SM_depthS %>%
  ggplot(aes(x = log_depth))+
  geom_hline(yintercept = 0, linetype = "11", col = "grey")+
  geom_ribbon(aes(ymin = l_89*100, ymax = u_89*100, fill = .category), alpha = .3)+
  geom_line(aes(y = median*100, col = .category), linewidth = 1, show.legend = F)+
   geom_point(data = dat_percL %>% 
               filter(size_class != ">100") %>% 
               mutate(.category = recode(type, 
                                    "feeding_frac"= "frac_feedingT",
                                    "investigating_circling_frac"= "frac_circ_invT",
                                    "passing_frac"= "frac_passingT"),
                      log_depth = log(depth)),
             aes(y = perc*100, col = .category))+
  facet_wrap(~.category)+
  theme_clean()
```

Impact of number of fish in frame

```{r}
epred_perc_SM_n_tot <-  epred_draws(
  fit_perc_SM,
  newdata = expand_grid(year = dat_model$year %>% unique(),
                        treatment = dat_model$treatment %>% unique(),
                        size_class = c("<50", "50-100"),
                        habitat = dat_model$habitat %>% unique(),
                        log_depth = mean(dat_model$depth) %>% 
                          log(),
                        log_n_frame_total = seq(min(dat_model$n_frame_total),
                                        max(dat_model$n_frame_total),
                                        length = 100) %>% 
                          log()),
  re_formula = NA)


epred_perc_SM_n_totS <- epred_perc_SM_n_tot %>% 
  group_by(.category, log_n_frame_total) %>% 
  summarise(median = median(.epred),
            l_66 = quantile(.epred, probs = (1-0.66)/2),
            u_66 = quantile(.epred, probs = 1 - (1-0.66)/2),
            l_89 = quantile(.epred, probs = (1-0.89)/2),
            u_89 = quantile(.epred, probs = 1-(1-0.89)/2),
            l95 = quantile(.epred, 0.025, na.rm = T),
            u95 = quantile(.epred, 0.975, na.rm = T),
      )

epred_perc_SM_n_totS %>%
  ggplot(aes(x = log_n_frame_total))+
  geom_hline(yintercept = 0, linetype = "11", col = "grey")+
  geom_ribbon(aes(ymin = l_89*100, ymax = u_89*100, fill = .category), alpha = .3)+
  geom_line(aes(y = median*100, col = .category), linewidth = 1, show.legend = F)+
   geom_point(data = dat_percL %>% 
               filter(size_class != ">100") %>% 
               mutate(.category = recode(type, 
                                    "feeding_frac"= "frac_feedingT",
                                    "investigating_circling_frac"= "frac_circ_invT",
                                    "passing_frac"= "frac_passingT"),
                      log_n_frame_total = log(n_frame_total )),
             aes(y = perc*100, col = .category))+
  facet_wrap(~.category)+
  theme_clean()
```

Impact of habitat

```{r}


epred_perc_SM %>%
  pivot_wider(names_from = .category, values_from = .epred) %>%
  ggtern(aes(x = frac_circ_invT, y = frac_passingT, z = frac_feedingT)) +
  geom_point(aes(col = habitat), alpha = .1, shape = 1) +
  stat_confidence_tern(
    aes(group = habitat),
    col = "white",
    breaks = c(0.89),
    show.legend = F
  ) +
  labs(
    x = "",
    xarrow  = "Circling/Investigating",
    y = "",
    yarrow  = "Passing",
    z = "",
    zarrow  = "Feeding"
  ) +
  guides(color = guide_legend(override.aes = list(alpha = 1)))+
  theme_tern_clean()
```

```{r}
epred_perc_SM %>%
  ggplot(aes(x = .epred, y = treatment)) +
  stat_slab(size = .3,
            aes(fill = habitat),
            slab_size = 0.5, 
            alpha = 0.3) +
  stat_slab(show.legend = F, 
            size = .3,
            fill = NA, 
            aes(colour = habitat), 
            slab_size = 0.5) +
  guides(color = guide_legend(override.aes = list(alpha = 1)),
         fill = guide_legend(override.aes = list(alpha = 1)))+
  facet_grid(year ~ size_class + .category)+
  theme_clean()
```

```{r}
epred_perc_SM %>%
  ggplot(aes(x = .epred, y = 1)) +
  stat_slab(size = .3,
            aes(fill = habitat),
            slab_size = 0.5, 
            alpha = 0.3) +
  stat_slab(show.legend = F, 
            size = .3,
            fill = NA, 
            aes(colour = habitat), 
            slab_size = 0.5) +
  guides(color = guide_legend(override.aes = list(alpha = 1)),
         fill = guide_legend(override.aes = list(alpha = 1)))+
  theme_clean()
```

#### Large fish

#### A less complex model has to be used because number of observations low

```{r}
fit_perc_L <- brm(
  cbind(frac_circ_invT, frac_passingT, frac_feedingT) ~
    factor(year) * treatment +
    s(log_n_frame_total) +
    habitat +
    s(log_depth) +
    (1 | video),
  data = dat_model %>%
    filter(size_class != ">100"),
  family =  dirichlet(),
  iter = 4000,
  warmup = 2000,
  chains = 4,
  cores = 4,
  seed = 123,
  control = list(max_treedepth = 12, adapt_delta = 0.98),
  file_refit = "on_change",
  threads = threading(2, static = T),
  file = here("models/fit_perc_L"),
  backend = "cmdstanr"
)
```

```{r}
summary(fit_perc_L)# 5 divergent transitions, Rhat OK
#plot(fit_perc_L) # OK
```

Get expected values

```{r}
epred_perc_L <-  epred_draws(
  fit_perc_L,
  newdata = expand_grid(year = dat_model$year %>% unique(),
                        treatment = dat_model$treatment %>% unique(),
                        habitat = dat_model$habitat %>% unique(),
                        log_depth = dat_model$log_depth %>% mean(),
                        log_n_frame_total = mean(dat_model$n_frame_total) %>% 
                          log()),
  re_formula = NA
) %>% 
  mutate(size_class = ">100")
```

Model Plots

```{r}
epred_perc_L %>%
  ggplot(aes(x = .epred, y = year)) +
  stat_slab(size = .3,
            aes(fill = treatment),
            slab_size = 0.5, 
            alpha = 0.3) +
  stat_slab(show.legend = F, 
            size = .3,
            fill = NA, 
            aes(colour = treatment), 
            slab_size = 0.5) +
  scale_colour_manual(values = cols_treat)+
  scale_fill_manual(values = cols_treat)+
  facet_grid(~ .category)
```

As tenary plot. white lines are 89% CIs

**Compare before after in one plot**

```{r}
epred_perc_L %>%
  pivot_wider(names_from = .category, values_from = .epred) %>%
  ggtern(aes(x = frac_circ_invT, y = frac_passingT, z = frac_feedingT)) +
  geom_point(aes(col = treatment), alpha = .1, shape = 1) +
  stat_confidence_tern(
    aes(group = treatment),
    col = "white",
    breaks = c(0.89),
    show.legend = F
  ) +
  labs(
    x = "",
    xarrow  = "Circling/Investigating",
    y = "",
    yarrow  = "Passing",
    z = "",
    zarrow  = "Feeding"
  ) +
  scale_colour_manual(values = cols_treat) +
  guides(color = guide_legend(override.aes = list(alpha = 1))) +
  facet_grid(~ year) +
  theme_tern_clean()
```

**Compare years in one plot**

```{r}
epred_perc_L %>%
  pivot_wider(names_from = .category, values_from = .epred) %>%
  ggtern(aes(x = frac_circ_invT, y = frac_passingT, z = frac_feedingT)) +
  geom_point(aes(col = year), alpha = .1, shape = 1) +
  stat_confidence_tern(
    aes(group = year),
    col = "white",
    breaks = c(0.89),
    show.legend = F
  ) +
  labs(
    x = "",
    xarrow  = "Circling/Investigating",
    y = "",
    yarrow  = "Passing",
    z = "",
    zarrow  = "Feeding"
  ) +
  scale_colour_manual(values = cols_year) +
  guides(color = guide_legend(override.aes = list(alpha = 1))) +
  facet_grid(~ treatment) +
  theme_tern_clean()
```

Estimated value for difference between treatments. If positive, more time spent with respective behavior after shark appeared.

```{r}
epred_perc_effect_L <- epred_perc_L %>%
  ungroup() %>%
  dplyr::select(year,
                treatment,
                .draw,
                .epred,
                .category,
                habitat,
                log_n_frame_total,
                log_depth) %>%
  pivot_wider(names_from = treatment, values_from = .epred) %>%
  mutate(diff = after - before)
 


epred_perc_effect_L %>%
  ggplot(aes(x = diff, y = year)) +
  geom_vline(xintercept = 0,
             linetype = "11",
             col = "grey") +
  stat_slab(alpha = 0.3, aes(fill = factor(year))) +
  stat_slab(
    show.legend = F,
    size = .5,
    fill = NA ,
    aes(col = factor(year))
  ) +
  scale_colour_manual(values = cols_year)+
  scale_fill_manual(values = cols_year)+
  facet_grid(~ .category, scales = "free") +
  theme_clean()
```

Compare distribution of expected values of effect to 0

```{r}
epred_perc_effect_L %>% 
  compare_0(variable = diff,
          year, .category)
```
