==> scripts/3_model_observationtime.qmd <==
---
title: "Model Total Observation Time Shark Present <> Absent"
author: "Andreas Eich"
format: html
editor: visual
---

### Libraries

```{r}
library(here) # easier paths, here() function starts in the folder of the RStudio project
library(tidyverse)
library(brms)
library(tidybayes)
library(ggtern)
library(kableExtra)
library(janitor)
library(bayesplot)
library(bayestestR)
library(emmeans)
library(marginaleffects)
library(htmlTable)

# get custom functions
source(here("scripts/functions.R"))
```

### Get data

```{r}
dat_obs <- readRDS(here("data/clean/clean_data.rds"))
```

Divide into small&medium and large data set

```{r}
dat_obs_SM <- dat_obs %>% 
  filter(size_class != "Large") %>% 
  droplevels()

dat_obs_L <- dat_obs %>% 
  filter(size_class == "Large") %>% 
  droplevels()
```

### Plot Raw Data

Plot raw data and mean ± SE of mean of mean observation time per video

```{r}
dat_obs %>%
  group_by(size_class, shark_in_video, year, video) %>%
  summarise(sec_total_video = mean(sec_total)) %>%
  group_by(size_class, shark_in_video, year) %>%
  summarise(
    mean_sec_total = mean(sec_total_video, na.rm = T),
    sd = sd(sec_total_video, na.rm = T),
    n = n(),
    se = ifelse(n >= 3, sd / sqrt(n), NA)
  ) %>%
  ggplot(aes(x = shark_in_video, col = factor(shark_in_video))) +
  geom_jitter(data = dat_obs, aes(y = sec_total), alpha = .2) +
  geom_errorbar(aes(
    ymin = mean_sec_total - se,
    ymax = mean_sec_total + se,
    col = factor(shark_in_video),
    width = .4
  )) +
  geom_point(aes(y = mean_sec_total)) +
  facet_grid(year ~ size_class , scales = "free") +
  theme_clean()
```

### Models

#### Small & medium fish

```{r}
fit_sec_total_SM <- brm(
  sec_total ~ year * shark_in_video * size_class +
    s(log_n_frame_total) +
    habitat +
    s(log_depth) +
    (1 | video),
  data = dat_obs_SM,
  family = Gamma(link = "log"),
  control = list(adapt_delta = 0.99, 
                 max_treedepth = 12),
  iter = 4000,
  warmup = 2000,
  chains = 4,
  cores = 4,
  seed = 123,
  file_refit = "on_change",
  file = here("models/fit_sec_total_SM"),
  threads = threading(2, static = T),
  backend = "cmdstanr"
)
```

Assess Model

```{r}
summary(fit_sec_total_SM)#36 divergent transitions, Rhat ok

ppc_dens_overlay(y = log1p(dat_obs_SM$sec_total),
                 yrep = log1p(posterior_predict(fit_sec_total_SM)[1:100, ]))

bayes_R2(fit_sec_total_SM)#0.49
```

Quickly plot results

```{r}
#conditional_effects(fit_sec_total_SM,plot = T, ask = F)
```

No impact

##### Plot Results for year shark_in_video size_class

Get Emmeans values

```{r}
epred_sec_total_SM <- emmeans(fit_sec_total_SM,
                              ~year * shark_in_video * size_class, 
                              typ = "response") %>%   
  gather_emmeans_draws() %>% 
  mutate(.value = exp(.value))# gather_emmeans_draws() removes back-transformation
```

Plot

```{r}

plot_observationtime_SM <- ggplot(data = epred_sec_total_SM, 
                                 aes(x = year, 
                                     col = paste(shark_in_video, size_class, sep = "_"))) +
  geom_point(
    data = dat_obs_SM,
    aes(y = sec_total),
    position =  position_jitterdodge(jitter.width = .2, dodge.width = .7),
    alpha = .1,
    size = 3,
    stroke = 0
  ) +
  stat_pointinterval(
    aes(y = .value),
    point_interval = mode_hdi,
    .width = c(.95, .89),
    position = position_dodge(width = .7),
    show.legend = F
  ) +
  scale_colour_manual(values = cols_sharks,
                      labels = c("absent", "absent", "present", "present")) +
  guides(color = guide_legend(override.aes = list(alpha = 1), title = "Shark")) +
  labs(x = NULL, y = "Observation time (sec)") +
  facet_grid( ~ size_class, scales = "free") +
  scale_y_continuous(trans = 'log10') +
  theme_clean(); plot_observationtime_SM
```

And save

```{r}
ggsave(filename = "observationtime_SM.pdf",
       plot = plot_observationtime_SM, 
       path = here("plots"),
       width = 15, height = 9, units = "cm")
```

##### Conditional effects for shark_in_video per year and size_class

```{r}
emmeans(fit_sec_total_SM, ~ shark_in_video | size_class + year, type = "response") %>%
  regrid() %>%
  pairs() %>%
  gather_emmeans_draws() %>%
  group_by(contrast, size_class, year) %>%
  median_hdi(.value) %>% 
  mutate(important = ifelse(.upper > 0 & .lower > 0 |
                       .upper < 0 & .lower < 0, "*","")) %>% 
  select(-.width, -.point, -.interval) %>% 
  mutate_if(is.numeric, txtRound,2) %>% 
  htmlTable(rnames = FALSE)

```

##### Conditional effects for habitat

```{r}
emmeans(fit_sec_total_SM, ~  habitat, type = "response") %>% plot()

emmeans(fit_sec_total_SM, ~  habitat, type = "response") %>%
  regrid() %>%
  pairs() %>%
  gather_emmeans_draws() %>%
  median_hdi(.value) %>% 
  mutate(important = ifelse(.upper > 0 & .lower > 0 |
                       .upper < 0 & .lower < 0, "*","")) %>% 
  select(-.width, -.point, -.interval) %>% 
  mutate_if(is.numeric, txtRound,2) %>% 
  htmlTable(rnames = FALSE)
```

No effect

##### Conditional effects for n fish in frame

```{r}
plot_epred_sec_total_SM_ntotal <- plot_predictions(
  fit_sec_total_SM,
  condition = c("log_n_frame_total"),
  conf_level = .95
) +
  geom_point(data = dat_obs_SM, aes(x = log_n_frame_total, y = sec_total)) +
  scale_y_continuous(trans = 'log10') +
  labs(x = "log(n fish in frame)", y = "Observation time (sec)") +
  theme_clean()


plot_epred_sec_total_SM_ntotal

ggsave(filename = "observationtime_SM_ntotal.pdf",
       plot = plot_epred_sec_total_SM_ntotal, 
       path = here("plots/SI"),
       width = 20, height = 15, units = "cm")


```

Slight increase

##### Conditional effects for depth

```{r}
plot_epred_sec_total_SM_depth <- plot_predictions(fit_sec_total_SM, 
                 condition = c("log_depth"), 
                 conf_level = .95)+
  geom_point(data = dat_obs_SM,
             aes(x = log_depth, y =sec_total))+
  scale_y_continuous(trans = 'log10') +
  labs(x = "log(depth)", y = "Observation time (sec)") +
  theme_clean()

plot_epred_sec_total_SM_depth

ggsave(filename = "observationtime_SM_depth.pdf",
       plot = plot_epred_sec_total_SM_depth, 
       path = here("plots/SI"),
       width = 20, height = 15, units = "cm")

```

No effect

#### Large fish

A less complex model has to be used because number of observations low. Smoothing cannot be used because number of covariate combinations is too low

```{r}
fit_sec_total_L <- brm(
  sec_total ~ year * shark_in_video +
    log_n_frame_total +
    habitat +
    log_depth +
    (1 | video),
  data = dat_obs_L,
  family = Gamma(link = "log"),
  control = list(max_treedepth = 12, adapt_delta = 0.99),
  iter = 4000,
  warmup = 2000,
  chains = 4,
  cores = 4,
  seed = 123,
  file_refit = "on_change",
  file = here("models/fit_sec_total_L"),
  threads = threading(2, static = T),
  backend = "cmdstanr"
)
```

```{r}
summary(fit_sec_total_L)#ok
pp_check(fit_sec_total_L, ndraws = 100)
bayes_R2(fit_sec_total_L)#0.45
```

##### Plot Results for year shark_in_video size_class

Get Emmeans values

```{r}
epred_sec_total_L <- emmeans(fit_sec_total_L,
                              ~year * shark_in_video, 
                              typ = "response") %>%   
  gather_emmeans_draws() %>% 
  mutate(.value = exp(.value))# gather_emmeans_draws() removes back-transformation
```

Plot

```{r}

plot_observationtime_L <- ggplot(data = epred_sec_total_L, 
                                 aes(x = year, 
                                     col = shark_in_video)) +
  geom_point(
    data = dat_obs_L,
    aes(y = sec_total),
    position =  position_jitterdodge(jitter.width = .2, dodge.width = .7),
    alpha = .1,
    size = 3,
    stroke = 0
  ) +
  stat_pointinterval(
    aes(y = .value),
    point_interval = mode_hdi,
    .width = c(.95, .89),
    position = position_dodge(width = .7),
    show.legend = F
  ) +
  scale_colour_manual(values = cols_sharks_L,
                      labels = c("absent", "present")) +
  guides(color = guide_legend(override.aes = list(alpha = 1), title = "Shark")) +
  labs(x = NULL, y = "Observation time (sec)") +
  scale_y_continuous(trans = 'log10') +
  theme_clean(); plot_observationtime_L
```

And save

```{r}
ggsave(filename = "observationtime_L.pdf",
       plot = plot_observationtime_L, 
       path = here("plots/SI"),
       width = 15/1.7, height = 9, units = "cm")
```

##### Conditional effects for shark_in_video per year

```{r}
emmeans(fit_sec_total_L, ~ shark_in_video | year, type = "response") %>%
  regrid() %>%
  pairs() %>%
  gather_emmeans_draws() %>%
  group_by(contrast, year) %>%
  # median_hdi returns 3 groups? => manually
  summarise(median = median(.value),
            bayestestR::hdi(.value)) %>% 
  rename(".lower"  = "CI_low",
         ".upper"  = "CI_high") %>% 
  select(-CI) %>% 
  mutate(important = ifelse(.upper > 0 & .lower > 0 |
                       .upper < 0 & .lower < 0, "*","")) %>% 
  mutate_if(is.numeric, txtRound,2) %>% 
  htmlTable(rnames = FALSE)

```

##### Conditional effects for habitat

```{r}
emmeans(fit_sec_total_L, ~  habitat, type = "response") %>% plot()

emmeans(fit_sec_total_L, ~  habitat, type = "response") %>%
  regrid() %>%
  pairs() %>%
  gather_emmeans_draws() %>%
  # median_hdi returns 3 groups? => manually
  summarise(median = median(.value),
            bayestestR::hdi(.value)) %>% 
  rename(".lower"  = "CI_low",
         ".upper"  = "CI_high") %>% 
  select(-CI) %>% 
  mutate(important = ifelse(.upper > 0 & .lower > 0 |
                       .upper < 0 & .lower < 0, "*","")) %>% 
  mutate_if(is.numeric, txtRound,2) %>% 
  htmlTable(rnames = FALSE)
```

No effect

##### Conditional effects for n fish in frame

```{r}
plot_epred_sec_total_L_ntotal <- plot_predictions(
  fit_sec_total_L,
  condition = c("log_n_frame_total"),
  conf_level = .95
) +
  geom_point(data = dat_obs_L, aes(x = log_n_frame_total, y = sec_total)) +
  scale_y_continuous(trans = 'log10') +
  labs(x = "log(n fish in frame)", y = "Observation time (sec)") +
  theme_clean()


plot_epred_sec_total_L_ntotal

ggsave(filename = "observationtime_L_ntotal.pdf",
       plot = plot_epred_sec_total_L_ntotal, 
       path = here("plots/SI"),
       width = 20, height = 15, units = "cm")


```

Slight increase

##### Conditional effects for depth

```{r}
plot_epred_sec_total_L_depth <- plot_predictions(fit_sec_total_L, 
                 condition = c("log_depth"), 
                 conf_level = .95)+
  geom_point(data = dat_obs_L,
             aes(x = log_depth, y =sec_total))+
  scale_y_continuous(trans = 'log10') +
  labs(x = "log(depth)", y = "Observation time (sec)") +
  theme_clean()

plot_epred_sec_total_L_depth

ggsave(filename = "observationtime_L_depth.pdf",
       plot = plot_epred_sec_total_L_depth, 
       path = here("plots/SI"),
       width = 20, height = 15, units = "cm")

```

No effect
