==> scripts/5_mod_boutsbites.qmd <==
---
title: "Model Bites and Bouts Present <> Absent"
author: "Andreas Eich"
format: html
editor: visual
---

### Libraries

```{r}
library(here) # easier paths, here() function starts in the folder of the RStudio project
library(tidyverse)
library(brms)
library(tidybayes)
library(ggtern)
library(kableExtra)
library(janitor)
library(bayesplot)
library(emmeans)
library(marginaleffects)
library(htmlTable)

# get custom functions
source(here("scripts/functions.R"))
```

### Get data

Filter for data with bites \> 0 and sec_feeding \> 0

```{r}
dat_boutbite <- readRDS(here("data/clean/clean_data.rds")) %>% 
  filter(bite > 0 & sec_feeding > 0)
```

Check any bout = 0

```{r}
dat_boutbite %>% 
  pull(bout) %>% 
  min()
```

OK

### Raw Data

#### Bites

Plot raw data and mean ± SE of mean of mean bouts during feeding observation per video

```{r}
dat_boutbite %>%
  group_by(size_class, shark_in_video, year, video) %>%
  summarise(mean_bites_per_video = mean(bite)) %>% 
  group_by(size_class, shark_in_video, year) %>%
  summarise(
    mean_bites = mean(mean_bites_per_video, na.rm = T),
    sd = sd(mean_bites_per_video, na.rm = T),
    n = n(),
    se = ifelse(n >= 3, sd / sqrt(n), NA)
  ) %>%
  ggplot() +
  geom_errorbar(aes(
    x = shark_in_video,
    ymin = mean_bites - se,
    ymax = mean_bites + se,
    col = factor(shark_in_video),
    width = .4
  )) +
  geom_point(aes(
    x = shark_in_video,
    y = mean_bites,
    col = factor(shark_in_video)
  )) +
  facet_grid(year ~ size_class , scales = "free") +
  theme_clean()
```

Only possible for small fish

#### Bouts

```{r}
dat_boutbite %>%
  group_by(size_class, shark_in_video, year, video) %>%
  summarise(mean_bouts_per_video = mean(bout)) %>% 
  group_by(size_class, shark_in_video, year) %>%
  summarise(
    mean_bouts = mean(mean_bouts_per_video, na.rm = T),
    sd = sd(mean_bouts_per_video, na.rm = T),
    n = n(),
    se = ifelse(n >= 3, sd / sqrt(n), NA)
  ) %>%
  ggplot() +
  geom_errorbar(aes(
    x = shark_in_video,
    ymin = mean_bouts - se,
    ymax = mean_bouts + se,
    col = factor(shark_in_video),
    width = .4
  )) +
  geom_point(aes(
    x = shark_in_video,
    y = mean_bouts,
    col = factor(shark_in_video)
  )) +
  facet_grid(year ~ size_class , scales = "free") +
  theme_clean()
```

Only possible for small fish

### Models Bites Small

Only Small Fish, very few observations, only assess year \* shark_in_video

```{r}
dat_boutbite_S <- dat_boutbite %>% 
  filter(size_class == "Small") %>% 
  select(-size_class)
```

```{r}
fit_bites_S <- brm(
  bite ~ year * shark_in_video + (1 | video),
  data = dat_boutbite_S,
  family = poisson(link = "log"),  
  control = list(max_treedepth = 16, adapt_delta = 0.99),
  iter = 4000,
  warmup = 2000,
  chains = 4,
  cores = 4,
  seed = 123,
  file_refit = "on_change",
  file = here("models/fit_bites_S"),
  threads = threading(2, static = T),
  backend = "cmdstanr"
)
```

Assess Model

```{r}
summary(fit_bites_S)#ok
ppc_dens_overlay(y = log1p(dat_boutbite_S$bite),
                 yrep = log1p(posterior_predict(fit_bites_S)[1:100, ]))
bayes_R2(fit_bites_S)#0.5500921
```

#### Plot Results for year shark_in_video size_class

Get Emmeans values

```{r}
epred_bites_S <- emmeans(fit_bites_S,
                              ~year * shark_in_video , 
                              typ = "response") %>%   
  gather_emmeans_draws() %>% 
  mutate(.value = exp(.value))# gather_emmeans_draws() removes back-transformation
```

Plot

```{r}

plot_bites_S <- ggplot(data = epred_bites_S %>% 
      mutate(size_class = "Small"), 
                                 aes(x = year, 
                                     col = shark_in_video)) +
  geom_point(
    data = dat_boutbite_S %>% 
      mutate(size_class = "Small"),
    aes(y = bite),
    position =  position_jitterdodge(jitter.width = .2, dodge.width = .7),
    alpha = .1,
    size = 3,
    stroke = 0
  ) +
  stat_pointinterval(
    aes(y = .value),
    point_interval = mode_hdi,
    .width = c(.95, .89),
    position = position_dodge(width = .7),
    show.legend = F
  ) +
  scale_colour_manual(values = cols_sharks_S) +
  guides(color = guide_legend(override.aes = list(alpha = 1), title = "Shark")) +
  labs(x = NULL, y = "Bites") +
  facet_grid( ~ size_class, scales = "free") +
  theme_clean(); plot_bites_S
```

And save to SI, only small effect

```{r}
ggsave(filename = "bites_S.pdf",
       plot = plot_bites_S, 
       path = here("plots/SI"),
       width = 15/1.7, height = 9, units = "cm")
```

##### Conditional effects for shark_in_video per year

```{r}
#plot
emmeans(fit_bites_S, ~ shark_in_video |  year, type = "response") %>%
  regrid() %>%
  pairs() %>% 
  plot()

#table
emmeans(fit_bites_S, ~ shark_in_video |  year, type = "response") %>%
  regrid() %>%
  pairs() %>%
  gather_emmeans_draws() %>%
  group_by(contrast, year) %>%
    # median_hdi returns 3 groups? => manually
  summarise(median = median(.value),
            bayestestR::hdi(.value)) %>% 
  rename(".lower"  = "CI_low",
         ".upper"  = "CI_high") %>% 
    select(-CI) %>% 
  mutate(important = ifelse(.upper > 0 & .lower > 0 |
                       .upper < 0 & .lower < 0, "*","")) %>% 
  mutate_if(is.numeric, txtRound,2) %>% 
  htmlTable(rnames = FALSE)

```

Very weak effect in 2016

### Models Bouts Small

Only Small Fish, very few observations, only assess year \* shark_in_video

```{r}
fit_bouts_S <- brm(
  bout ~ year * shark_in_video + (1 | video),
  data = dat_boutbite_S,
  family = poisson(link = "log"),  
  control = list(max_treedepth = 16, adapt_delta = 0.99),
  iter = 4000,
  warmup = 2000,
  chains = 4,
  cores = 4,
  seed = 123,
  file_refit = "on_change",
  file = here("models/fit_bouts_S"),
  threads = threading(2, static = T),
  backend = "cmdstanr"
)
```

Assess Model

```{r}
summary(fit_bouts_S)#ok
ppc_dens_overlay(y = log1p(dat_boutbite_S$bout),
                 yrep = log1p(posterior_predict(fit_bouts_S)[1:100, ]))
bayes_R2(fit_bouts_S)#0.800257
```

#### Plot Results for year shark_in_video size_class

Get Emmeans values

```{r}
epred_bouts_S <- emmeans(fit_bouts_S,
                              ~year * shark_in_video , 
                              typ = "response") %>%   
  gather_emmeans_draws() %>% 
  mutate(.value = exp(.value))# gather_emmeans_draws() removes back-transformation
```

Plot

```{r}
plot_bouts_S <- ggplot(data = epred_bouts_S %>% 
      mutate(size_class = "Small"), 
                                 aes(x = year, 
                                     col = shark_in_video)) +
  geom_point(
    data = dat_boutbite_S %>% 
      mutate(size_class = "Small"),
    aes(y = bout),
    position =  position_jitterdodge(jitter.width = .2, dodge.width = .7),
    alpha = .1,
    size = 3,
    stroke = 0
  ) +
  stat_pointinterval(
    aes(y = .value),
    point_interval = mode_hdi,
    .width = c(.95, .89),
    position = position_dodge(width = .7),
    show.legend = F
  ) +
  scale_colour_manual(values = cols_sharks_S) +
  guides(color = guide_legend(override.aes = list(alpha = 1), title = "Shark")) +
  labs(x = NULL, y = "Bouts") +
  facet_grid( ~ size_class, scales = "free") +
  theme_clean(); plot_bouts_S
```

And save as pdf

```{r}
# ggsave(filename = "bouts_S.pdf",
#        plot = plot_bouts_S, 
#        path = here("plots"),
#        width = 15/1.7, height = 9, units = "cm")
```

To combine with boutrate, save as rds w/o legend

```{r}
saveRDS(plot_bouts_S+
          theme(legend.position = "none"), here("plots/plot_bouts_S.rds"))
```

##### Conditional effects for shark_in_video per year

```{r}
#plot
emmeans(fit_bouts_S, ~ shark_in_video |  year, type = "response") %>%
  regrid() %>%
  pairs() %>% 
  plot()

#table
emmeans(fit_bouts_S, ~ shark_in_video |  year, type = "response") %>%
  regrid() %>%
  pairs() %>%
  gather_emmeans_draws() %>%
  group_by(contrast, year) %>%
    # median_hdi returns 3 groups? => manually
  summarise(median = median(.value),
            bayestestR::hdi(.value)) %>% 
  rename(".lower"  = "CI_low",
         ".upper"  = "CI_high") %>% 
    select(-CI) %>% 
  mutate(important = ifelse(.upper > 0 & .lower > 0 |
                       .upper < 0 & .lower < 0, "*","")) %>% 
  mutate_if(is.numeric, txtRound,2) %>% 
  htmlTable(rnames = FALSE)

```

Effect in 2016
